requirements.txt
cudatoolkit>=11.0.221
cudnn>=8.0.5
imageio>=2.9.0
numpy>=1.19.2
pandas>=1.2.1
pillow>=8.1.0
pip>=21.0.1
pytorch>=1.7.1
scikit-image>=0.18.1
scikit-learn>=0.24.1
scipy>=1.5.2
torchvision>=0.8.2
tqdm>=4.56.0
